+++ packages/themes/imove/src/Theme/Theme.ExpandListItem.js
@@ -1,0 +1,40 @@
+import { css } from "styled-components"
+import { ExpandListItem, theming } from "@staccx/base"
+
+export const ExpandButton = theming.createVariants(
+  {
+    [theming.VARIANT_DEFAULT]: css`
+      padding: ${theming.spacing.small} 0;
+      font-size: ${theming.font("small")};
+      font-weight: ${theming.fontWeight.bold};
+      color: ${theming.color.primary};
+
+      :focus,
+      :hover {
+        > svg {
+          fill: ${theming.color.black} !important;
+        }
+      }
+    `
+  },
+  ExpandListItem.themeProps.listButton
+)
+
+export const ExpandIcon = theming.createVariants(
+  {
+    [theming.VARIANT_DEFAULT]: css`
+      right: 0;
+      fill: ${theming.color.primary};
+    `
+  },
+  ExpandListItem.themeProps.listIcon
+)
+
+export const ExpandedItem = theming.createVariants(
+  {
+    [theming.VARIANT_DEFAULT]: css`
+      padding: 0;
+    `
+  },
+  ExpandListItem.themeProps.expandedItem
+)