+++ packages/apps/bank-in-a-box/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [8.71.0](https://bitbucket.org/stacc-flow/bento/compare/v8.70.1...v8.71.0) (2019-03-05)
+
+**Note:** Version bump only for package @staccx/bank-in-a-box
+
+
+
+
+
+## [8.70.1](https://bitbucket.org/stacc-flow/bento/compare/v8.70.0...v8.70.1) (2019-03-05)
+
+
+### Bug Fixes
+
+* **biab:** fix i18n and onclick for documents ([399c1d3](https://bitbucket.org/stacc-flow/bento/commits/399c1d3))
+
+
+
+
+
+# [8.70.0](https://bitbucket.org/stacc-flow/bento/compare/v8.69.1...v8.70.0) (2019-03-01)
+
+**Note:** Version bump only for package @staccx/bank-in-a-box
+
+
+
+
+
+## [8.69.1](https://bitbucket.org/stacc-flow/bento/compare/v8.69.0...v8.69.1) (2019-03-01)
+
+
+### Bug Fixes
+
+* **biab:** fix withdraw ([c4324cf](https://bitbucket.org/stacc-flow/bento/commits/c4324cf))
+
+
+
+
+
+# [8.69.0](https://bitbucket.org/stacc-flow/bento/compare/v8.68.1...v8.69.0) (2019-03-01)
+
+
+### Bug Fixes
+
+* **biab:** fix endaccount ([5a4a8e8](https://bitbucket.org/stacc-flow/bento/commits/5a4a8e8))
+
+
+
+
+
 # [8.68.0](https://bitbucket.org/stacc-flow/bento/compare/v8.67.0...v8.68.0) (2019-02-28)