+++ packages/internal/base/CHANGELOG.md
@@ -3,6 +3,58 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [8.71.0](https://bitbucket.org/stacc-flow/bento/compare/v8.70.1...v8.71.0) (2019-03-05)
+
+**Note:** Version bump only for package @staccx/base
+
+
+
+
+
+## [8.70.1](https://bitbucket.org/stacc-flow/bento/compare/v8.70.0...v8.70.1) (2019-03-05)
+
+
+### Bug Fixes
+
+* **base:** fix missing theme prop ([dacff35](https://bitbucket.org/stacc-flow/bento/commits/dacff35))
+
+
+
+
+
+# [8.70.0](https://bitbucket.org/stacc-flow/bento/compare/v8.69.1...v8.70.0) (2019-03-01)
+
+
+### Features
+
+* **base:** new component development ([2c25521](https://bitbucket.org/stacc-flow/bento/commits/2c25521))
+
+
+
+
+
+## [8.69.1](https://bitbucket.org/stacc-flow/bento/compare/v8.69.0...v8.69.1) (2019-03-01)
+
+
+### Bug Fixes
+
+* **base:** form traversechildren ([c3b42fd](https://bitbucket.org/stacc-flow/bento/commits/c3b42fd))
+
+
+
+
+
+# [8.69.0](https://bitbucket.org/stacc-flow/bento/compare/v8.68.1...v8.69.0) (2019-03-01)
+
+
+### Bug Fixes
+
+* **biab:** fix endaccount ([5a4a8e8](https://bitbucket.org/stacc-flow/bento/commits/5a4a8e8))
+
+
+
+
+
 # [8.68.0](https://bitbucket.org/stacc-flow/bento/compare/v8.67.0...v8.68.0) (2019-02-28)